+++ src/test/java/seedu/address/ui/AppParametersTest.java
@@ -1,4 +1,4 @@
-package seedu.address.ui.guitests;
+package seedu.address.ui;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
