+++ src/test/java/seedu/address/ui/GuiUnitTest.java
@@ -1,4 +1,4 @@
-package seedu.address.ui.guitests;
+package seedu.address.ui;
 
 import java.util.Optional;
