+++ src/test/java/seedu/address/ui/HelpWindowTest.java
@@ -1,4 +1,4 @@
-package seedu.address.ui.guitests;
+package seedu.address.ui;
 
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
@@ -11,7 +11,6 @@ import org.testfx.api.FxToolkit;
 
 import guitests.guihandles.HelpWindowHandle;
 import javafx.stage.Stage;
-import seedu.address.ui.HelpWindow;
 
 public class HelpWindowTest extends GuiUnitTest {
