+++ src/test/java/seedu/address/ui/UiPartTest.java
@@ -12,7 +12,6 @@ import org.junit.jupiter.api.io.TempDir;
 
 import javafx.fxml.FXML;
 import seedu.address.MainApp;
-import seedu.address.ui.guitests.TestFxmlObject;
 
 public class UiPartTest {
